* 삭제부분 반환 값 설정 완료

+++ src/main/java/com/example/demo/feed/service/FeedService.java
@@ -2,11 +2,21 @@ package com.example.demo.feed.service;
 
 
 import com.example.demo.feed.domain.*;
+import com.example.demo.friend.domain.FriendShipRepository;
+import com.example.demo.friend.domain.FriendShipState;
+import com.example.demo.friend.domain.Friendship;
+import com.example.demo.friend.domain.PersonId;
 import com.example.demo.user.domain.User;
+import com.example.demo.user.service.UserService;
 import com.example.demo.user.service.UserSessionService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.client.HttpClientErrorException;
+
+import java.net.http.HttpResponse;
 import java.nio.file.AccessDeniedException;
 
 @Service
@@ -14,12 +24,28 @@ import java.nio.file.AccessDeniedException;
 public class FeedService {
     private final FeedRepository feedRepository;
     private final UserSessionService userSessionService;
-
+    private final FriendShipRepository friendShipRepository;
+    private final UserService userService;
 
     @Transactional
-    public void deleteFeedByFeedId(Long id) {
+    public RequestResult deleteFeedByFeedId(Long id) throws AccessDeniedException {
         Feed feed = feedRepository.findFeedByIdAndDeleted(id, false);
+        PersonId myId = PersonId.create(Long.parseLong(userService.findUserInfo()[0]));
+        PersonId friendId = PersonId.create(feed.getId());
+        Friendship friendship;
+
+        if ((friendship=friendShipRepository.findFriendshipByFriender_FrienderIdAndFriendee_FriendeeId(myId, friendId))==null){
+            return RequestResult.BAD_REQUEST;
+        }
+
+        if (myId.getId().equals(friendId.getId())){
+
+        }else if (friendship.getFriendState()!= FriendShipState.ACCEPT){
+            return RequestResult.UNAUTHORIZED;
+        }
+
         feed.delete();
+        return RequestResult.OK;
     }
